pyproject.toml
[project]
name = "mcp-client-gradio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.5.1",
    "gradio>=5.31.0",
    "ipython>=9.2.0",
    "magika>=0.6.2",
    "mcp-server-fetch>=2025.1.17",
    "ollama>=0.5.1",
    "pyyaml>=6.0.2",
]
